more precision in pageboxes
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/branches/experimental@6959 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/image/pdftoepdf.c
@@ -611,7 +611,7 @@ void read_pdf_info(image_dict * idict)
     ppint rotate = 0;
     int pdf_major_version_found = 1;
     int pdf_minor_version_found = 3;
-    float xsize, ysize, xorig, yorig;
+    double xsize, ysize, xorig, yorig;
     if (img_type(idict) == IMG_TYPE_PDF) {
         pdf_doc = refPdfDocument(img_filepath(idict), FE_FAIL, img_userpassword(idict), img_ownerpassword(idict));
     } else if (img_type(idict) == IMG_TYPE_PDFMEMSTREAM) {
